Implement solution to problem 8.

+++ src/ProjectEuler/Problem5.java
@@ -18,9 +18,9 @@ class Problem5 {
     }
 
     private static int getGreatestCommonDivisor(int number1, int number2) {
-        int remainder;
+        byte remainder;
         while (number2 > 0) {
-            remainder = number1 % number2;
+            remainder = (byte) (number1 % number2);
             number1 = number2;
             number2 = remainder;
         }
